post construct

+++ src/main/java/com/ll/sb20231114/domain/article/article/controller/ArticleController.java
@@ -10,17 +10,21 @@ import org.springframework.web.bind.annotation.PostMapping;
 
 import com.ll.sb20231114.domain.article.article.entity.Article;
 import com.ll.sb20231114.domain.article.article.service.ArticleService;
+import com.ll.sb20231114.domain.member.member.entity.Member;
+import com.ll.sb20231114.domain.member.member.service.MemberService;
 import com.ll.sb20231114.global.rq.Rq;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 
 @Controller
 @RequiredArgsConstructor
 public class ArticleController {
 	private final ArticleService articleService;
+	private final MemberService memberService;
 	private final Rq rq;
 
 	@GetMapping("/article/list")
@@ -43,6 +47,9 @@ public class ArticleController {
 
 	@GetMapping("/article/write")
 	String showWrite() {
+		if (!rq.isLogined())
+			throw new RuntimeException("로그인 후 이용해주세요.");
+
 		return "article/article/write";
 	}
 
@@ -55,14 +62,22 @@ public class ArticleController {
 	}
 
 	@PostMapping("/article/write")
+	@SneakyThrows
 	String write(@Valid WriteForm writeForm) {
-		Article article = articleService.write(writeForm.title, writeForm.body);
+		if (!rq.isLogined())
+			throw new RuntimeException("로그인 후 이용해주세요.");
+
+		Member loginedMember = rq.getMember();
+		Article article = articleService.write(loginedMember, writeForm.title, writeForm.body);
 
 		return rq.redirect("/article/list", "%d번 게시물 생성되었습니다.".formatted(article.getId()));
 	}
 
 	@GetMapping("/article/modify/{id}")
 	String showModify(Model model, @PathVariable long id) {
+		if (!rq.isLogined())
+			throw new RuntimeException("로그인 후 이용해주세요.");
+
 		Article article = articleService.findById(id).get();
 
 		model.addAttribute("article", article);
@@ -80,6 +95,9 @@ public class ArticleController {
 
 	@PostMapping("/article/modify/{id}")
 	String modify(@PathVariable long id, @Valid ModifyForm modifyForm) {
+		if (!rq.isLogined())
+			throw new RuntimeException("로그인 후 이용해주세요.");
+
 		articleService.modify(id, modifyForm.title, modifyForm.body);
 
 		return rq.redirect("/article/list", "%d번 게시물 수정되었습니다.".formatted(id));
@@ -87,6 +105,9 @@ public class ArticleController {
 
 	@GetMapping("/article/delete/{id}")
 	String delete(@PathVariable long id) {
+		if (!rq.isLogined())
+			throw new RuntimeException("로그인 후 이용해주세요.");
+
 		articleService.delete(id);
 
 		return rq.redirect("/article/list", "%d번 게시물 삭제되었습니다.".formatted(id));
